Update diffusers version and set cuda().half() to fit GPU

+++ scripts/txt2img.py
@@ -60,7 +60,9 @@ def load_model_from_config(config, ckpt, verbose=False):
         print("unexpected keys:")
         print(u)
 
-    model.cuda()
+    # model.cuda()
+    # halfsized to fit in 8GB RTX1070
+    model.cuda().half()
     model.eval()
     return model
